Split gp, mk -> gp, mk, mk_details header files
