+++ src/backend/executor/nodeMotion.c
@@ -25,8 +25,8 @@
 #include "parser/parsetree.h"
 #include "utils/lsyscache.h"
 #include "utils/tuplesort.h"
-#include "utils/tuplesort_gp.h"
 #include "utils/tuplesort_mk.h"
+#include "utils/tuplesort_mk_details.h"
 #include "miscadmin.h"
 #include "nodes/makefuncs.h"
 #include "utils/memutils.h"
