+++ src/backend/utils/sort/tuplesort_mk.c
@@ -117,10 +117,10 @@
 #include "utils/pg_rusage.h"
 #include "utils/syscache.h"
 #include "utils/tuplesort.h"
-#include "utils/tuplesort_gp.h"
 #include "utils/pg_locale.h"
 #include "utils/builtins.h"
 #include "utils/tuplesort_mk.h"
+#include "utils/tuplesort_mk_details.h"
 #include "utils/string_wrapper.h"
 #include "utils/faultinjector.h"
