+++ src/backend/utils/sort/tuplesort_mkheap.c
@@ -7,8 +7,8 @@
 #include "postgres.h"
 #include "access/genam.h"
 #include "utils/tuplesort.h"
-#include "utils/tuplesort_gp.h"
 #include "utils/tuplesort_mk.h"
+#include "utils/tuplesort_mk_details.h"
 #include "utils/datum.h"
 #include "utils/builtins.h"
 #include "miscadmin.h"
