+++ src/backend/utils/sort/tuplesort_mkqsort.c
@@ -13,8 +13,8 @@
 #include "postgres.h"
 #include "access/genam.h"
 #include "utils/tuplesort.h"
-#include "utils/tuplesort_gp.h"
 #include "utils/tuplesort_mk.h"
+#include "utils/tuplesort_mk_details.h"
 
 #include "miscadmin.h"
